requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
Pillow==9.2.0
pyTelegramBotAPI==4.6.0
requests==2.28.1
urllib3==1.26.10
